test(api): prove compaction-after-compaction lineage end to end (#57)
New compactions.e2e-spec drives a real double compaction over HTTP
(threshold=1, fake model): the second row chains to the first
(parentId, monotonic uptoSeq), its model input is previous summary +
delta only — never the full history — and the next chat turn's prompt
leads with the latest summary and excludes everything absorbed. The
fake streaming model client moves to test/support.ts for reuse.
