+++ apps/api/test/chats-messages.e2e-spec.ts
@@ -10,11 +10,6 @@
 import { INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import request from 'supertest';
-import {
-  type LanguageModelUsage,
-  type ModelMessage,
-  type streamText,
-} from 'ai';
 import { AppModule } from './../src/app.module';
 import { configureApp } from './../src/app.setup';
 import { type Message } from './../src/db/schema';
@@ -23,11 +18,9 @@ import {
   ChatsRepository,
   MessagesRepository,
 } from './../src/chats/chats-repository';
-import { MissingModelCredentialError } from './../src/models/model-client';
 import { ModelsService } from './../src/models/models.service';
-import { TITLE_SYSTEM_PROMPT } from './../src/chats/title';
 import { turnTelemetryLogger } from './../src/chats/turn-telemetry';
-import { cookieOf, streamedText } from './support';
+import { FakeModelsService, cookieOf, streamedText } from './support';
 
 const hasDb = !!process.env.POSTGRES_URL;
 const d = hasDb ? describe : describe.skip;
@@ -50,169 +43,6 @@ async function waitFor(
       throw new Error('Timed out waiting for condition');
     }
     await new Promise((resolve) => setTimeout(resolve, 5));
-  }
-}
-
-type FakeTurn = {
-  messages: ModelMessage[];
-  abortSignal?: AbortSignal;
-  aborted: boolean;
-};
-
-class FakeStreamingModelClient {
-  readonly turns: FakeTurn[] = [];
-  // Title-generation calls (#78) are tracked separately: they are async post-turn
-  // work, so counting them in `turns` would make every chat-turn assertion racy.
-  readonly titleTurns: ModelMessage[][] = [];
-  titleResponse: string | Promise<string> = 'Generated Title';
-  readonly model = 'gpt-4o-mini';
-  readonly provider = 'openai';
-  responses: string[] = ['fake assistant'];
-  usage: LanguageModelUsage = {
-    inputTokens: 3,
-    inputTokenDetails: {
-      noCacheTokens: 1,
-      cacheReadTokens: 2,
-      cacheWriteTokens: 0,
-    },
-    cachedInputTokens: 2,
-    outputTokens: 5,
-    outputTokenDetails: { textTokens: 4, reasoningTokens: 1 },
-    totalTokens: 8,
-    reasoningTokens: 1,
-  };
-  shouldFinish = true;
-  delayMs = 0;
-
-  streamText(input: {
-    system?: string;
-    messages: ModelMessage[];
-    abortSignal?: AbortSignal;
-    onFinish?: (event: {
-      text: string;
-      usage: LanguageModelUsage;
-      finishReason: string;
-    }) => void | Promise<void>;
-    onError?: (event: { error: unknown }) => void | Promise<void>;
-  }): ReturnType<typeof streamText> {
-    if (input.system === TITLE_SYSTEM_PROMPT) {
-      this.titleTurns.push(input.messages);
-      return {
-        text: Promise.resolve(this.titleResponse),
-      } as unknown as ReturnType<typeof streamText>;
-    }
-
-    const response =
-      this.responses[this.turns.length] ?? this.responses[0] ?? '';
-    const turn: FakeTurn = {
-      messages: input.messages,
-      abortSignal: input.abortSignal,
-      aborted: false,
-    };
-    this.turns.push(turn);
-
-    input.abortSignal?.addEventListener('abort', () => {
-      turn.aborted = true;
-    });
-
-    const stream = new ReadableStream({
-      start: async (controller) => {
-        controller.enqueue({
-          type: 'start',
-          messageId: `fake-${this.turns.length}`,
-        });
-        controller.enqueue({ type: 'text-start', id: 'text-1' });
-
-        if (this.delayMs > 0) {
-          await new Promise((resolve) => setTimeout(resolve, this.delayMs));
-        }
-
-        if (input.abortSignal?.aborted) {
-          turn.aborted = true;
-          const error = new Error('aborted');
-          await input.onError?.({ error });
-          controller.error(error);
-          return;
-        }
-
-        controller.enqueue({
-          type: 'text-delta',
-          id: 'text-1',
-          delta: response,
-        });
-        controller.enqueue({ type: 'text-end', id: 'text-1' });
-
-        if (input.abortSignal?.aborted) {
-          turn.aborted = true;
-          const error = new Error('aborted');
-          await input.onError?.({ error });
-          controller.error(error);
-          return;
-        }
-
-        if (this.shouldFinish) {
-          await input.onFinish?.({
-            text: response,
-            usage: this.usage,
-            finishReason: 'stop',
-          });
-          controller.enqueue({ type: 'finish' });
-        }
-
-        controller.close();
-      },
-    });
-
-    const toResponse = () => {
-      const sse = stream.pipeThrough(
-        new TransformStream({
-          transform(part, controller) {
-            controller.enqueue(`data: ${JSON.stringify(part)}\n\n`);
-          },
-          flush(controller) {
-            controller.enqueue('data: [DONE]\n\n');
-          },
-        }),
-      );
-      return new Response(sse.pipeThrough(new TextEncoderStream()), {
-        headers: {
-          'content-type': 'text/event-stream',
-          'cache-control': 'no-cache',
-          connection: 'keep-alive',
-          'x-vercel-ai-ui-message-stream': 'v1',
-        },
-      });
-    };
-
-    return {
-      text: Promise.resolve(response),
-      textStream: new ReadableStream({
-        start(controller) {
-          controller.enqueue(response);
-          controller.close();
-        },
-      }) as never,
-      fullStream: new ReadableStream() as never,
-      consumeStream: async () => {},
-      toUIMessageStreamResponse: toResponse,
-    } as unknown as ReturnType<typeof streamText>;
-  }
-}
-
-class FakeModelsService {
-  credential: string | null = 'sk-test';
-  readonly client = new FakeStreamingModelClient();
-
-  resolveModelCredential(userId: string): string {
-    if (!this.credential) {
-      throw new MissingModelCredentialError(userId);
-    }
-
-    return this.credential;
-  }
-
-  createOpenAIClient() {
-    return this.client;
   }
 }
