+++ apps/api/test/support.ts
@@ -1,10 +1,15 @@
 /**
- * Shared e2e test helpers. The session-cookie format and the AI SDK SSE event
- * shape are protocol facts each spec used to restate — keep them in one place
- * so a change (cookie name, stream event schema) can't silently miss a copy.
+ * Shared e2e test helpers. The session-cookie format, the AI SDK SSE event
+ * shape, and the fake streaming model client are protocol facts each spec used
+ * to restate — keep them in one place so a change (cookie name, stream event
+ * schema, fake behavior) can't silently miss a copy.
  */
 
 import type request from 'supertest';
+import type { LanguageModelUsage, ModelMessage, streamText } from 'ai';
+
+import { TITLE_SYSTEM_PROMPT } from './../src/chats/title';
+import { MissingModelCredentialError } from './../src/models/model-client';
 
 /** Extracts the llame session cookie pair from a response, or '' when absent. */
 export const cookieOf = (res: request.Response): string => {
@@ -47,4 +52,167 @@ export function streamedText(body: string): string {
     )
     .map((event) => event.delta)
     .join('');
+}
+
+export type FakeTurn = {
+  messages: ModelMessage[];
+  abortSignal?: AbortSignal;
+  aborted: boolean;
+};
+
+export class FakeStreamingModelClient {
+  readonly turns: FakeTurn[] = [];
+  // Title-generation calls (#78) are tracked separately: they are async post-turn
+  // work, so counting them in `turns` would make every chat-turn assertion racy.
+  readonly titleTurns: ModelMessage[][] = [];
+  titleResponse: string | Promise<string> = 'Generated Title';
+  readonly model = 'gpt-4o-mini';
+  readonly provider = 'openai';
+  responses: string[] = ['fake assistant'];
+  usage: LanguageModelUsage = {
+    inputTokens: 3,
+    inputTokenDetails: {
+      noCacheTokens: 1,
+      cacheReadTokens: 2,
+      cacheWriteTokens: 0,
+    },
+    cachedInputTokens: 2,
+    outputTokens: 5,
+    outputTokenDetails: { textTokens: 4, reasoningTokens: 1 },
+    totalTokens: 8,
+    reasoningTokens: 1,
+  };
+  shouldFinish = true;
+  delayMs = 0;
+
+  streamText(input: {
+    system?: string;
+    messages: ModelMessage[];
+    abortSignal?: AbortSignal;
+    onFinish?: (event: {
+      text: string;
+      usage: LanguageModelUsage;
+      finishReason: string;
+    }) => void | Promise<void>;
+    onError?: (event: { error: unknown }) => void | Promise<void>;
+  }): ReturnType<typeof streamText> {
+    if (input.system === TITLE_SYSTEM_PROMPT) {
+      this.titleTurns.push(input.messages);
+      return {
+        text: Promise.resolve(this.titleResponse),
+      } as unknown as ReturnType<typeof streamText>;
+    }
+
+    const response =
+      this.responses[this.turns.length] ?? this.responses[0] ?? '';
+    const turn: FakeTurn = {
+      messages: input.messages,
+      abortSignal: input.abortSignal,
+      aborted: false,
+    };
+    this.turns.push(turn);
+
+    input.abortSignal?.addEventListener('abort', () => {
+      turn.aborted = true;
+    });
+
+    const stream = new ReadableStream({
+      start: async (controller) => {
+        controller.enqueue({
+          type: 'start',
+          messageId: `fake-${this.turns.length}`,
+        });
+        controller.enqueue({ type: 'text-start', id: 'text-1' });
+
+        if (this.delayMs > 0) {
+          await new Promise((resolve) => setTimeout(resolve, this.delayMs));
+        }
+
+        if (input.abortSignal?.aborted) {
+          turn.aborted = true;
+          const error = new Error('aborted');
+          await input.onError?.({ error });
+          controller.error(error);
+          return;
+        }
+
+        controller.enqueue({
+          type: 'text-delta',
+          id: 'text-1',
+          delta: response,
+        });
+        controller.enqueue({ type: 'text-end', id: 'text-1' });
+
+        if (input.abortSignal?.aborted) {
+          turn.aborted = true;
+          const error = new Error('aborted');
+          await input.onError?.({ error });
+          controller.error(error);
+          return;
+        }
+
+        if (this.shouldFinish) {
+          await input.onFinish?.({
+            text: response,
+            usage: this.usage,
+            finishReason: 'stop',
+          });
+          controller.enqueue({ type: 'finish' });
+        }
+
+        controller.close();
+      },
+    });
+
+    const toResponse = () => {
+      const sse = stream.pipeThrough(
+        new TransformStream({
+          transform(part, controller) {
+            controller.enqueue(`data: ${JSON.stringify(part)}\n\n`);
+          },
+          flush(controller) {
+            controller.enqueue('data: [DONE]\n\n');
+          },
+        }),
+      );
+      return new Response(sse.pipeThrough(new TextEncoderStream()), {
+        headers: {
+          'content-type': 'text/event-stream',
+          'cache-control': 'no-cache',
+          connection: 'keep-alive',
+          'x-vercel-ai-ui-message-stream': 'v1',
+        },
+      });
+    };
+
+    return {
+      text: Promise.resolve(response),
+      textStream: new ReadableStream({
+        start(controller) {
+          controller.enqueue(response);
+          controller.close();
+        },
+      }) as never,
+      fullStream: new ReadableStream() as never,
+      consumeStream: async () => {},
+      toUIMessageStreamResponse: toResponse,
+    } as unknown as ReturnType<typeof streamText>;
+  }
+}
+
+export class FakeModelsService {
+  credential: string | null = 'sk-test';
+  readonly client = new FakeStreamingModelClient();
+
+  resolveModelCredential(userId: string): string {
+    if (!this.credential) {
+      throw new MissingModelCredentialError(userId);
+    }
+
+    return this.credential;
+  }
+
+  createOpenAIClient() {
+    return this.client;
+  }
 }
